+++ modules/mysql/main.tf
@@ -25,23 +25,30 @@ locals {
 
   databases = { for db in var.additional_databases : db.name => db }
   users     = { for u in var.additional_users : u.name => u }
+
+  // HA method using REGIONAL availability_type requires binary logs to be enabled
+  binary_log_enabled = var.availability_type == "REGIONAL" ? true : lookup(var.backup_configuration, "binary_log_enabled", null)
+  backups_enabled    = var.availability_type == "REGIONAL" ? true : lookup(var.backup_configuration, "enabled", null)
 }
 
 resource "google_sql_database_instance" "default" {
-  project          = var.project_id
-  name             = var.name
-  database_version = var.database_version
-  region           = var.region
+  provider            = google-beta
+  project             = var.project_id
+  name                = var.name
+  database_version    = var.database_version
+  region              = var.region
+  encryption_key_name = var.encryption_key_name
 
   settings {
     tier                        = var.tier
     activation_policy           = var.activation_policy
+    availability_type           = var.availability_type
     authorized_gae_applications = var.authorized_gae_applications
     dynamic "backup_configuration" {
       for_each = [var.backup_configuration]
       content {
-        binary_log_enabled = lookup(backup_configuration.value, "binary_log_enabled", null)
-        enabled            = lookup(backup_configuration.value, "enabled", null)
+        binary_log_enabled = local.binary_log_enabled
+        enabled            = local.backups_enabled
         start_time         = lookup(backup_configuration.value, "start_time", null)
       }
     }