+++ modules/postgresql/main.tf
@@ -27,10 +27,12 @@ locals {
 }
 
 resource "google_sql_database_instance" "default" {
-  project          = var.project_id
-  name             = var.name
-  database_version = var.database_version
-  region           = var.region
+  provider            = google-beta
+  project             = var.project_id
+  name                = var.name
+  database_version    = var.database_version
+  region              = var.region
+  encryption_key_name = var.encryption_key_name
 
   settings {
     tier                        = var.tier
@@ -40,7 +42,7 @@ resource "google_sql_database_instance" "default" {
     dynamic "backup_configuration" {
       for_each = [var.backup_configuration]
       content {
-        binary_log_enabled = lookup(backup_configuration.value, "binary_log_enabled", null)
+        binary_log_enabled = false
         enabled            = lookup(backup_configuration.value, "enabled", null)
         start_time         = lookup(backup_configuration.value, "start_time", null)
       }