+++ test/integration/mysql-ha/controls/mysql.rb
@@ -17,13 +17,14 @@ basename   = attribute('name')
 authorized_network = attribute('authorized_network')
 
 describe google_sql_database_instances(project: project_id).where(instance_name: /#{basename}/) do
-  its(:count) { should eq 5 }
+  its(:count) { should eq 4 }
 end
 
 describe google_sql_database_instance(project: project_id, database: basename) do
   let(:expected_settings) {
     {
       activation_policy: "ALWAYS",
+      availability_type: "REGIONAL",
       data_disk_size_gb: 10,
       data_disk_type: "PD_SSD",
       kind: "sql#settings",
@@ -55,41 +56,6 @@ describe google_sql_database_instance(project: project_id, database: basename) d
   it { expect(location_preference).to include(kind: "sql#locationPreference", zone: "us-central1-c") }
   it { expect(maintenance_window).to include(kind: "sql#maintenanceWindow", day: 7, hour: 12, update_track: "stable") }
   it { expect(user_labels).to include(foo: "bar") }
-end
-
-describe google_sql_database_instance(project: project_id, database: "#{basename}-failover-test") do
-  let(:expected_settings) {
-    {
-      activation_policy: "ALWAYS",
-      data_disk_size_gb: 10,
-      data_disk_type: "PD_SSD",
-      kind: "sql#settings",
-      pricing_plan: "PER_USE",
-      replication_type: "SYNCHRONOUS",
-      storage_auto_resize: true,
-      storage_auto_resize_limit: 0,
-      tier: "db-n1-standard-1",
-    }
-  }
-  let(:settings)                { subject.settings.item }
-  let(:ip_configuration)        { settings[:ip_configuration] }
-  let(:database_flags)          { settings[:database_flags] }
-  let(:location_preference)     { settings[:location_preference] }
-  let(:maintenance_window)      { settings[:maintenance_window] }
-  let(:user_labels)             { settings[:user_labels] }
-
-  its(:backend_type)     { should eq 'SECOND_GEN' }
-  its(:database_version) { should eq 'MYSQL_5_7' }
-  its(:state)            { should eq 'RUNNABLE' }
-  its(:region)           { should eq 'us-central1' }
-  its(:gce_zone)         { should eq 'us-central1-a' }
-
-  it { expect(settings).to include(expected_settings) }
-  it { expect(ip_configuration).to include(authorized_networks: [{kind: 'sql#aclEntry', name: "#{project_id}-cidr", value: authorized_network}], ipv4_enabled: true, require_ssl: false) }
-  it { expect(database_flags).to include(name: "long_query_time", value: "1") }
-  it { expect(location_preference).to include(kind: "sql#locationPreference", zone: "us-central1-a") }
-  it { expect(maintenance_window).to include(kind: "sql#maintenanceWindow", day: 3, hour: 20, update_track: "canary") }
-  it { expect(user_labels).to include(baz: "boo") }
 end
 
 %i[a b c].each_with_index do |zone, index|